% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_test_funcs.R
\name{dummy_child_func}
\alias{dummy_child_func}
\title{A dummy child function that we will "hotfix" in tests}
\usage{
dummy_child_func(x)
}
\description{
A dummy child function that we will "hotfix" in tests
}
