% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dummy_parent_func}
\alias{dummy_parent_func}
\title{A dummy parent function to test namespace injection}
\usage{
dummy_parent_func(x)
}
\arguments{
\item{x}{Input value to pass to child function}
}
\value{
A character string with parent output
}
\description{
A dummy parent function to test namespace injection
}
\examples{
# Call the parent function which internally calls dummy_child_func
result <- hotpatchR:::dummy_parent_func("sample")
print(result)

}
