% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotpatchR.R
\name{inject_patch}
\alias{inject_patch}
\title{Inject a runtime patch into a locked package namespace}
\usage{
inject_patch(pkg, patch_list, lock = TRUE)
}
\arguments{
\item{pkg}{A package name as a string.}

\item{patch_list}{A named list of functions to overwrite in the package namespace.}

\item{lock}{Whether to re-lock bindings after patching.}
}
\value{
Invisibly TRUE on success.
}
\description{
Inject a runtime patch into a locked package namespace
}
