% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotpatch.R
\name{test_patched_dir}
\alias{test_patched_dir}
\title{Run testthat tests against a patched package namespace}
\usage{
test_patched_dir(pkg, test_path = "tests/testthat", reporter = "summary")
}
\arguments{
\item{pkg}{Package name that has been patched in memory.}

\item{test_path}{Path to tests to run.}

\item{reporter}{testthat reporter name or object.}
}
\value{
Result object from testthat::test_dir.
}
\description{
Run testthat tests against a patched package namespace
}
\examples{
# Inject a patch to the package
inject_patch(
  pkg = "hotpatchR",
  patch_list = list(dummy_child_func = function(x) {
    paste("PATCHED! Input:", x)
  })
)

# Run tests against the patched package
# (Note: this requires tests to exist in tests/testthat directory)
# test_patched_dir(pkg = "hotpatchR")
}
