% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dummy_child_func}
\alias{dummy_child_func}
\title{A dummy child function that we will "hotfix" in tests}
\usage{
dummy_child_func(x)
}
\arguments{
\item{x}{Input value}
}
\value{
A character string with child output
}
\description{
A dummy child function that we will "hotfix" in tests
}
\examples{
# This function is called by dummy_parent_func
# It serves as an example of a function that can be hotpatched
result <- dummy_child_func("example")
print(result)

}
