% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_test_funcs.R
\name{dummy_parent_func}
\alias{dummy_parent_func}
\title{A dummy parent function to test namespace injection}
\usage{
dummy_parent_func(x)
}
\description{
A dummy parent function to test namespace injection
}
