% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotpatchR.R
\name{undo_patch}
\alias{undo_patch}
\title{Undo a previously injected patch}
\usage{
undo_patch(pkg, names = NULL)
}
\arguments{
\item{pkg}{Package name or environment.}

\item{names}{Character vector of patched object names to restore. If NULL, restore all stored backups for pkg.}
}
\value{
Invisibly TRUE on success.
}
\description{
Undo a previously injected patch
}
